requirements.txt
click
paramiko
requests
python-slugify
